Migrate to `symfony/flex`; get as far as commands working in this world
This includes moving migrations from an automatic post-connect event to an explicit command, which I think is a
more "usual" approach. This is partly driven by finding that modern Doctrine Events does not seem to include the
global connection completing as an event that one should be listening for. Ideally we would also "migrate" to
Doctrine Migrations but I suggest we consider this after this already quite large set of changes.