Give Twig a `/tmp` location to cache in
For now, work on the assumption that any cache is better than none! It seems
like Twig bundle only supports filesystem caching, and it's not particularly
valuable to share this cache between ECS tasks. The key thing is that there is
one and that there's minimal configuration mess to think about.

Fixes #368